Refactor: move methods to draw.cpp
