+++ character_making.cpp
@@ -4,6 +4,7 @@
 #include "character.hpp"
 #include "class.hpp"
 #include "config.hpp"
+#include "draw.hpp"
 #include "i18n.hpp"
 #include "input.hpp"
 #include "macro.hpp"
