+++ tcg.cpp
@@ -1,6 +1,7 @@
 #include "audio.hpp"
 #include "config.hpp"
 #include "ctrl_file.hpp"
+#include "draw.hpp"
 #include "elona.hpp"
 #include "filesystem.hpp"
 #include "input.hpp"
